+++ app/test/search/handlers_test.dart
@@ -49,7 +49,7 @@ void main() {
           'packages': [
             {
               'package': 'pkg_foo',
-              'score': closeTo(0.33, 0.01),
+              'score': closeTo(0.32, 0.01),
             }
           ],
         });
@@ -63,7 +63,7 @@ void main() {
           'packages': [
             {
               'package': 'pkg_foo',
-              'score': closeTo(0.30, 0.01),
+              'score': closeTo(0.28, 0.01),
             }
           ],
         });
@@ -78,7 +78,7 @@ void main() {
               'packages': [
                 {
                   'package': 'pkg_foo',
-                  'score': closeTo(0.30, 0.01),
+                  'score': closeTo(0.28, 0.01),
                 }
               ],
             });